TG-11: add postman collection, fixes in controller

+++ src/controllers/tasks.js
@@ -1,23 +1,39 @@
 import Tasks from '../models/Tasks';
 
+const { ObjectId } = require('mongoose').Types;
+
+const isValidObjectId = (id) => {
+  if (ObjectId.isValid(id)) {
+    if ((String)(new ObjectId(id)) === id) { return true; }
+    return false;
+  }
+  return false;
+};
+
 const deleteTask = async (req, res) => {
+  const { id } = req.params;
   try {
-    const { id } = req.params;
+    if (!isValidObjectId(id)) {
+      return res.status(400).json({
+        message: `Invalid id ${id}`,
+        error: true,
+      });
+    }
     const taskFound = await Tasks.findByIdAndDelete(id);
     if (!taskFound) {
       return res.status(400).json({
         message: `The task with the id ${id} was not found`,
-        success: true,
+        error: true,
       });
     }
     return res.status(204).json({
       message: `Task with id ${id} deleted succesfully`,
-      success: true,
       data: taskFound,
+      error: false,
     });
   } catch (error) {
     return res.status(500).json({
-      success: false,
+      error: true,
     });
   }
 };
@@ -26,27 +42,31 @@ const updateTask = async (req, res) => {
   try {
     const { id } = req.params;
     const updates = req.body;
-    console.log(updates);
+    if (!isValidObjectId(id)) {
+      return res.status(400).json({
+        message: `Invalid id ${id}`,
+        error: true,
+      });
+    }
     const taskFound = await Tasks.findByIdAndUpdate(
       id,
       updates,
       { new: true },
     );
     if (!taskFound) {
-      res.status(404).json({
+      return res.status(404).json({
         message: `Couldn't find task with id ${id}`,
-        success: false,
-      });
-    } else {
-      res.status(200).json({
-        message: `Modified task with id ${id}`,
-        success: true,
-        data: taskFound,
+        error: true,
       });
     }
+    return res.status(200).json({
+      message: `Modified task with id ${id}`,
+      data: taskFound,
+      error: false,
+    });
   } catch (error) {
-    res.status(500).json({
-      success: false,
+    return res.status(500).json({
+      error: true,
     });
   }
 };
